+++ spec/lib/dry/validation/matchers/validate_matcher_spec.rb
@@ -21,7 +21,7 @@ module Dry::Validation::Matchers
           required(:pets).filled(:array?)
           required(:other).filled(:hash?)
           optional(:hair_color).filled(:str?, included_in?: %w(blue orange))
-          optional(:address).value(max_size?: 10)
+          optional(:address).value(min_size?: 2, max_size?: 10)
         end
       end
     end
@@ -204,6 +204,20 @@ module Dry::Validation::Matchers
           value(included_in: %w(orange blue white))
         expect(matcher.matches?(schema_class)).to be false
       end
+    end
+
+    it "checks value against `min_size`" do
+      matcher = described_class.new(:address, :optional).
+        value(min_size: 3)
+      expect(matcher.matches?(schema_class)).to be false
+
+      matcher = described_class.new(:address, :optional).
+        value(min_size: 2)
+      expect(matcher.matches?(schema_class)).to be true
+
+      matcher = described_class.new(:address, :optional).
+        value(min_size: 1)
+      expect(matcher.matches?(schema_class)).to be false
     end
 
     it "checks value against `max_size`" do