+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [Unreleased]
+### Added
+- `min_size` validation
+
 ## [1.0.0] - 2019-04-04
 ### Added
 - Remove (I know, under "added") activesupport requirement